split off sage_jupyter.py
